projects/DMML#06/src/train_model.py
"""
train_model.py — Zero-Day Phishing URL Detector
Trains Logistic Regression + Constrained Random Forest.
Generates confusion matrix, ROC curve, feature importance plots.
"""

import os
import pandas as pd
import numpy as np
import matplotlib
matplotlib.use("Agg")
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.model_selection import train_test_split, cross_val_score, StratifiedKFold
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import (
    accuracy_score, precision_score, recall_score, f1_score,
    classification_report, confusion_matrix, roc_curve, auc,
    roc_auc_score
)
import joblib

OUT_DIR = os.path.dirname(os.path.abspath(__file__))

# ── 1. Load data ───────────────────────────────────────────────────
data = pd.read_csv("../2. feature_extraction/url_features.csv")
X = data.drop("label", axis=1)
y = data["label"]

print(f"Dataset  : {len(data)} samples")
print(f"Phishing : {y.sum()} | Legitimate: {(y==0).sum()}")
print(f"Features : {X.shape[1]} → {list(X.columns)}\n")

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=42, stratify=y
)
print(f"Train: {len(X_train)}   Test: {len(X_test)}\n")

cv = StratifiedKFold(n_splits=5, shuffle=True, random_state=42)

# ── 2. Logistic Regression ─────────────────────────────────────────
print("=" * 55)
print("LOGISTIC REGRESSION")
print("=" * 55)
lr = LogisticRegression(max_iter=2000, random_state=42)
lr.fit(X_train, y_train)
lr_pred  = lr.predict(X_test)
lr_proba = lr.predict_proba(X_test)[:, 1]

print(f"Accuracy  : {accuracy_score(y_test, lr_pred):.4f}")
print(f"Precision : {precision_score(y_test, lr_pred):.4f}")
print(f"Recall    : {recall_score(y_test, lr_pred):.4f}")
print(f"F1-Score  : {f1_score(y_test, lr_pred):.4f}")
print(f"AUC-ROC   : {roc_auc_score(y_test, lr_proba):.4f}")
print(classification_report(y_test, lr_pred))
lr_cv = cross_val_score(lr, X, y, cv=cv, scoring="accuracy")
print(f"5-Fold CV : {lr_cv.mean():.4f} ± {lr_cv.std():.4f}\n")

# ── 3. Random Forest (constrained — realistic accuracy) ────────────
print("=" * 55)
print("RANDOM FOREST (constrained)")
print("=" * 55)

# Constraints prevent memorisation of easy structural differences
# and produce realistic accuracy (93-96%) vs inflated 99%+
rf = RandomForestClassifier(
    n_estimators=100,
    max_depth=20,           # prevents overfitting on path_length
    min_samples_leaf=5,     # requires evidence before predicting
    min_samples_split=10,
    max_features="sqrt",    # standard generalisation setting
    random_state=42,
    n_jobs=-1
)
rf.fit(X_train, y_train)
rf_pred  = rf.predict(X_test)
rf_proba = rf.predict_proba(X_test)[:, 1]

print(f"Accuracy  : {accuracy_score(y_test, rf_pred):.4f}")
print(f"Precision : {precision_score(y_test, rf_pred):.4f}")
print(f"Recall    : {recall_score(y_test, rf_pred):.4f}")
print(f"F1-Score  : {f1_score(y_test, rf_pred):.4f}")
print(f"AUC-ROC   : {roc_auc_score(y_test, rf_proba):.4f}")
print(classification_report(y_test, rf_pred))
rf_cv = cross_val_score(rf, X, y, cv=cv, scoring="accuracy")
print(f"5-Fold CV : {rf_cv.mean():.4f} ± {rf_cv.std():.4f}\n")

# ── 4. Save model ──────────────────────────────────────────────────
joblib.dump(rf, os.path.join(OUT_DIR, "phishing_model.pkl"))
print("Model saved: phishing_model.pkl\n")

# ── 5. Confusion matrices ──────────────────────────────────────────
fig, axes = plt.subplots(1, 2, figsize=(12, 5))
fig.suptitle("Confusion Matrices — Zero-Day Phishing URL Detector",
             fontsize=14, fontweight="bold")

for ax, pred, title in zip(
    axes,
    [lr_pred, rf_pred],
    ["Logistic Regression", "Random Forest (saved model)"]
):
    cm = confusion_matrix(y_test, pred)
    sns.heatmap(cm, annot=True, fmt="d", cmap="Blues", ax=ax,
                xticklabels=["Legitimate", "Phishing"],
                yticklabels=["Legitimate", "Phishing"],
                linewidths=0.5, linecolor="white")
    ax.set_title(
        f"{title}\nAcc: {accuracy_score(y_test,pred):.3f}  "
        f"F1: {f1_score(y_test,pred):.3f}", fontsize=11)
    ax.set_xlabel("Predicted label", fontsize=10)
    ax.set_ylabel("True label", fontsize=10)

plt.tight_layout()
plt.savefig(os.path.join(OUT_DIR, "confusion_matrices.png"), dpi=150, bbox_inches="tight")
plt.close()
print("Saved: confusion_matrices.png")

# ── 6. ROC curve ──────────────────────────────────────────────────
fig, ax = plt.subplots(figsize=(8, 6))
for proba, label, color in [
    (lr_proba, "Logistic Regression", "#E8593C"),
    (rf_proba, "Random Forest",       "#185FA5"),
]:
    fpr, tpr, _ = roc_curve(y_test, proba)
    ax.plot(fpr, tpr, color=color, lw=2,
            label=f"{label}  (AUC = {auc(fpr,tpr):.4f})")

ax.plot([0,1],[0,1],"k--",lw=1,label="Random baseline (AUC = 0.50)")
ax.set_xlim([0.0,1.0]); ax.set_ylim([0.0,1.02])
ax.set_xlabel("False Positive Rate", fontsize=12)
ax.set_ylabel("True Positive Rate", fontsize=12)
ax.set_title("ROC Curve — Zero-Day Phishing URL Detector",
             fontsize=13, fontweight="bold")
ax.legend(loc="lower right", fontsize=11)
ax.grid(alpha=0.3)
plt.savefig(os.path.join(OUT_DIR, "roc_curve.png"), dpi=150, bbox_inches="tight")
plt.close()
print("Saved: roc_curve.png")

# ── 7. Feature importance ──────────────────────────────────────────
importances = pd.Series(
    rf.feature_importances_, index=X.columns).sort_values(ascending=True)

fig, ax = plt.subplots(figsize=(9, 6))
colors = ["#185FA5" if v >= importances.quantile(0.75) else "#9FE1CB"
          for v in importances]
importances.plot(kind="barh", ax=ax, color=colors, edgecolor="white")
ax.set_title("Feature Importance — Random Forest (Bias-Corrected Dataset)",
             fontsize=13, fontweight="bold")
ax.set_xlabel("Mean decrease in impurity", fontsize=11)
ax.grid(axis="x", alpha=0.3)
for i, v in enumerate(importances):
    ax.text(v+0.001, i, f"{v:.4f}", va="center", fontsize=8)
plt.tight_layout()
plt.savefig(os.path.join(OUT_DIR, "feature_importance.png"), dpi=150,
            bbox_inches="tight")
plt.close()
print("Saved: feature_importance.png")

# ── 8. Summary ────────────────────────────────────────────────────
print("\n" + "="*55)
print("SUMMARY (5-Fold Stratified CV)")
print("="*55)
print(pd.DataFrame({
    "Model":           ["Logistic Regression", "Random Forest"],
    "CV Mean Acc":     [f"{lr_cv.mean():.4f}", f"{rf_cv.mean():.4f}"],
    "CV Std":          [f"±{lr_cv.std():.4f}", f"±{rf_cv.std():.4f}"],
    "Test Accuracy":   [f"{accuracy_score(y_test,lr_pred):.4f}",
                        f"{accuracy_score(y_test,rf_pred):.4f}"],
    "AUC-ROC":         [f"{roc_auc_score(y_test,lr_proba):.4f}",
                        f"{roc_auc_score(y_test,rf_proba):.4f}"],
}).to_string(index=False))
print(f"\nAll outputs saved to: {OUT_DIR}")