projects/DMML#06/src/requirements.txt
streamlit
scikit-learn
pandas
numpy
joblib
python-Levenshtein
matplotlib
seaborn